+++ src/main/resources/dot.png
